changes made on the menu page
