+++ src/models/items/TagsModel.ts
@@ -1,0 +1,47 @@
+import mongoose, { Document, Model, Schema } from "mongoose";
+import { CommonDBInterface } from "../../lib/interfaces/DBinterfaces";
+
+export interface ITag extends Document, CommonDBInterface {
+  name: string;
+  is_active: boolean;
+  is_deleted: boolean;
+  created_at: Date;
+  updated_at: Date;
+}
+
+const TagSchema: Schema<ITag> = new Schema<ITag>(
+  {
+    name: {
+      type: String,
+      required: true,
+      unique: true,
+      trim: true,
+      lowercase: true,
+    },
+    is_active: {
+      type: Boolean,
+      default: true,
+    },
+    is_deleted: {
+      type: Boolean,
+      default: false,
+    },
+    created_at: {
+      type: Date,
+      default: Date.now,
+    },
+    updated_at: {
+      type: Date,
+      default: Date.now,
+    },
+  },
+  {
+    timestamps: {
+      createdAt: "created_at",
+      updatedAt: "updated_at",
+    },
+  }
+);
+
+const Tag: Model<ITag> = mongoose.model<ITag>("Tag", TagSchema);
+export default Tag;
